+++ include/samurai/storage/std/algebraic_array.hpp
